requirements.txt
pandas==2.0.2
requests==2.26.0
boto3==1.18.12
python-dotenv
mysql-connector-python
